Add Collection and some commands

+++ app/src/main/kotlin/proga_lab_5/city/Coordinates.kt
@@ -1,6 +1,6 @@
 package proga_lab_5.city
 
-class Coordinates(private var x: Long, private var y: Long) {
+class Coordinates(private var x: Long, private var y: Float) {
 
     fun getX(): Long = x
 
@@ -8,9 +8,9 @@ class Coordinates(private var x: Long, private var y: Long) {
         this.x = x
     }
 
-    fun getY(): Long = y
+    fun getY(): Float = y
 
-    fun setY(y: Long){
+    fun setY(y: Float){
         this.y = y
     }
